+++ tmux/all.sh
@@ -1,4 +1,17 @@
-#!/bin/sh
+#!/bin/bash
+set -e
 
-echo "Ensuring tmux install"
-brew install tmux
+if test $PLATFORM == "osx"; then
+  result=`/usr/local/bin/brew --prefix tmux`
+  if test -e "$result" ; then
+    echo "tmux installed"
+  else
+    echo "install tmux with brew"
+    brew install tmux
+  fi
+elif test "$PLATFORM" == "linux"; then
+  exit 1;
+else
+  exit 1;
+fi
+
